pyproject.toml
[project]
name = "simcrane"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "matplotlib>=3.10.8",
    "mujoco>=3.4.0",
    "numpy>=2.4.1",
    "pyyaml>=6.0.3",
    "zmq>=0.0.0",
]
